requirements.txt
fastapi
uvicorn
ollama
python-multipart
faster-whisper 
pypdf
pydub